requirements.txt
# RAG Agent — Dépendances Python
# Installer avec : pip install -r requirements.txt

# ── LangChain core ────────────────────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0

# ── Google Generative AI ──────────────────────────────────────────────────────
google-generativeai>=0.5.0

# ── Vector store ──────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── Document loaders ──────────────────────────────────────────────────────────
pypdf>=4.0.0          # Lecture PDF
unstructured>=0.13.0  # Loader avancé (optionnel)

# ── Env & config ──────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tiktoken>=0.7.0       # Comptage tokens (optionnel mais recommandé)
